requirements.txt
aiogrpc==1.8
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
base58==2.1.1
bech32==1.2.0
bitarray==2.9.2
bitstring==4.2.3
bolt11==2.0.6
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.6
daphne==4.1.2
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
grpcio-tools==1.62.1
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jwcrypto==1.5.6
lndgrpc==0.2.0
lnurl==0.1.1
msgpack==1.0.8
mypy-protobuf==3.5.0
oauthlib==3.2.2
packaging==24.0
pillow==10.4.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.15
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rsa==4.9
secp256k1==0.14.0
segno==1.6.1
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
types-protobuf==4.24.0.20240311
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
utils==1.0.2
uvicorn==0.29.0
websockets==12.0
whitenoise==6.6.0
zope.interface==6.3
